completed scripts and now can be run using automated chrooting, plus updated readme with instructions

+++ build.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 REPODIR="$(pwd)"
-echo $REPODIR
 BINARYNAME="steemd"
 WORKROOT="$HOME/calibrd-work"
 if [[ ! -d $WORKROOT ]]; then
@@ -8,6 +7,7 @@ if [[ ! -d $WORKROOT ]]; then
 fi
 
 # update submodules every time to be sure repo is always ready to build
+cd $REPODIR
 git submodule update --init --recursive
 
 # Colours for console output
@@ -36,14 +36,11 @@ function cleanup {
   
   prstat "Cleaning up..."
   
-  # Remove the chroot if it has been created
-  if [ $MOUNTEDSYSFOLDERS ]; then
-    prstat "Unmounting chroot bind mounts"
-    sudo umount $WORKROOT/ubuntu14/dev/pts
-    sudo umount $WORKROOT/ubuntu14/dev
-    sudo umount $WORKROOT/ubuntu14/proc
-    sudo umount $WORKROOT/ubuntu14/sys
-  fi
+  prstat "Unmounting chroot bind mounts"
+  sudo umount $WORKROOT/ubuntu14/dev/pts
+  sudo umount $WORKROOT/ubuntu14/dev
+  sudo umount $WORKROOT/ubuntu14/proc
+  sudo umount $WORKROOT/ubuntu14/sys
 
   # Each step of the process has an initialising indication and
   # a completed indicator. Only initialised but incomplete steps
@@ -54,9 +51,6 @@ function cleanup {
     prstat "removing $WORKROOT"
     sudo rm -rf $WORKROOT
   fi
-
-  # Let's blow this popstand!
-  exit
 }
 
 
@@ -81,6 +75,7 @@ prstat "Checking for necessary prerequisites..."
 install_pkg 'devscripts'
 install_pkg 'debootstrap'
 install_pkg 'pbuilder'
+install_pkg 'gnupg2'
 
 # Create work directory if it does not exist
 if [[ ! -d $WORKROOT ]]; then
@@ -92,7 +87,7 @@ prstat "Entering $WORKROOT..."
 cd $WORKROOT
 
 # Check if base image has been created
-if [[ ! -f $WORKROOT/.ubuntu14.tgz ]]; then
+if [[ ! -f $WORKROOT/.ubuntu14.tgz || ! -f $WORKROOT/ubuntu14.tgz ]]; then
   # Clean up mess if previous attempt was interrupted
   if [ -f $WORKROOT/ubuntu14.tgz ]; then
     prfalse "Base image creation was interrupted, cleaning up"
@@ -127,6 +122,9 @@ if [[ ! -f $WORKROOT/ubuntu14/.complete ]]; then
   sudo tar zxfp ../ubuntu14.tgz
   cd ..
   
+  # if this step needed to be repeated, then so do all subsequent
+  rm $WORKROOT/.cmake $WORKROOT/.appimage $WORKROOT/.boost
+
   # Process complete, mark it complete
   touch $WORKROOT/ubuntu14/.complete
   prtrue "Ubuntu 14.04 base image unpacked successfully"
@@ -135,11 +133,17 @@ else
   prtrue "Ubuntu 14.04 base image was already unpacked"
 fi
 
-# Check if Cmake was downloaded yet
-if [[ ! -f $WORKROOT/.cmake ]]; then
+# Check if Cmake download was marked completed, was downloaded yet, or correctly downloaded
+if [[ ! -f $WORKROOT/.cmake || ! -f $WORKROOT/cmake-3.2.2.tar.gz || \
+  ! $( sha256sum $WORKROOT/cmake-3.2.2.tar.gz | cut -d ' ' -f 1 ) == \
+  "ade94e6e36038774565f2aed8866415443444fb7a362eb0ea5096e40d5407c78" ]]; then
   # Download was not completed, clean up
   if [[ -f $WORKROOT/cmake-3.2.2.tar.gz ]]; then
-    # Delete intrerrupted download
+    # If file is not correct, delete intrerrupted download
+    if [[ ! $( sha256sum $WORKROOT/cmake-3.2.2.tar.gz | cut -d ' ' -f 1 ) == \
+      "ade94e6e36038774565f2aed8866415443444fb7a362eb0ea5096e40d5407c78" ]]; then
+      prfalse "Cmake download was corrupted, retry build script"
+    fi
     rm -f $WORKROOT/cmake-3.2.2.tar.gz
     prfalse "Removed incomplete download of Cmake 3.2.2"
   fi
@@ -151,8 +155,11 @@ if [[ ! -f $WORKROOT/.cmake ]]; then
     cp cmake-3.2.2.tar.gz $WORKROOT/ubuntu14/
 
     # TODO: check binary was not corrupted
-    prfalse "TODO: Check that cmake 3.2.2 source tarball was not corrupt"
-
+    if [[ ! $( sha256sum $WORKROOT/cmake-3.2.2.tar.gz | cut -d ' ' -f 1 ) == \
+      "ade94e6e36038774565f2aed8866415443444fb7a362eb0ea5096e40d5407c78" ]]; then
+      prfalse "Cmake download was corrupted, retry build script"
+      exit 1
+    fi
     # Process complete, mark it complete
     touch $WORKROOT/.cmake
     prtrue "Cmake 3.2.2 was downloaded and copied into chroot"
@@ -163,19 +170,13 @@ else
 fi
 
 # Check if AppImageKit was downloaded and copied into chroot
-# and that the AppImage folder skeleton was copied over
-if [[ ! -f $WORKROOT/.appimage ]]; then
+# and that the AppImage folder skeleton was created
+if [[ ! -f $WORKROOT/.appimage || ! -d $WORKROOT/$BINARYNAME-x86_64.AppDir || ! -d $WORKROOT/cli_wallet-x86_64.AppDir ]]; then
   # Clean up mess if previous attempt was interrupted
   if [[ -f $WORKROOT/appimagetool-x86_64.AppImage ]]; then
     # remove previous incomplete download
     rm -f $WORKROOT/appimagetool-x86_64.AppImage
     prfalse "Removed incomplete download of AppImage tool"
-  fi
-  # Clean up previous AppImage skeleton copy if it was interrupted
-  if [[ ! -d $WORKROOT/calibrd/calibrd.AppDir ]]; then
-    # Remove previous incomplete folder
-    rm -rf $WORKROOT/calibrd/calibrd.AppDir
-    prtrue "Removed improperly copied AppDir skeleton"
   fi
 
   # Download AppImageKit
@@ -187,10 +188,32 @@ if [[ ! -f $WORKROOT/.appimage ]]; then
   # Copy into chroot
   cp appimagetool-x86_64.AppImage $WORKROOT/ubuntu14/
 
-  # Copy AppDir skeleton into chroot
-  prstat "Creating AppDir skeleton in chroot"
-  mkdir -p $WORKROOT/calibrd/calibrd.AppDir/usr/bin
-  mkdir -p $WORKROOT/calibrd/calibrd.AppDir/usr/lib
+  # Create AppDir skeleton in chroot
+  prstat "Creating AppDir skeleton in chroot for $BINARYNAME"
+  mkdir -p $WORKROOT/$BINARYNAME-x86_64.AppDir/usr/bin
+  mkdir -p $WORKROOT/$BINARYNAME-x86_64.AppDir/usr/lib
+  mkdir -p $WORKROOT/$BINARYNAME-x86_64.AppDir/lib
+  echo "[Desktop Entry]">$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  echo "Name=$BINARYNAME">>$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  echo "Exec=$BINARYNAME">>$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  echo "Comment=$BINARYNAME server">>$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  echo "Icon=$BINARYNAME">>$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  echo "Type=server">>$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  echo "Categories=server blockchain">>$WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.desktop
+  touch $WORKROOT/$BINARYNAME-x86_64.AppDir/$BINARYNAME.png
+
+  prstat "Creating AppDir skeleton in chroot for cli_wallet"
+  mkdir -p $WORKROOT/cli_wallet-x86_64.AppDir/usr/bin
+  mkdir -p $WORKROOT/cli_wallet-x86_64.AppDir/usr/lib
+  mkdir -p $WORKROOT/cli_wallet-x86_64.AppDir/lib
+  echo "[Desktop Entry]">$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  echo "Name=cli_wallet">>$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  echo "Exec=cli_wallet">>$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  echo "Comment=cli_wallet for $BINARYNAME">>$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  echo "Icon=cli_wallet">>$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  echo "Type=cli app">>$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  echo "Categories=client blockchain">>$WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.desktop
+  touch $WORKROOT/cli_wallet-x86_64.AppDir/cli_wallet.png
 
   # Process completed successfully, mark complete
   touch $WORKROOT/.appimage
@@ -200,8 +223,8 @@ else
   prtrue "AppImageKit already downloaded and copied into chroot"
 fi
 
-# Check if boost download was already done
-if [ ! -f $WORKROOT/.boost ]; then
+# Check if boost download was already done, or if the source is missing from the chroot
+if [[ ! -f $WORKROOT/.boost || ! -f $WORKROOT/ubuntu14/boost_1_60_0.tar.bz2 ]]; then
   # Remove incomplete download if it was started and not finished
   if [[ -f $WORKROOT/boost_1_60_0.tar.bz2 ]]; then
     # Delete incomplete download
@@ -209,18 +232,27 @@ if [ ! -f $WORKROOT/.boost ]; then
     prfalse "Removed incomplete download of Boost source"
   fi
 
-  # Download Boost
-  prstat "Downloading boost 1.60..."
-  URL='http://sourceforge.net/projects/boost/files/boost/1.60.0/boost_1_60_0.tar.bz2/download'
-  wget -c "$URL" -O $WORKROOT/boost_1_60_0.tar.bz2 &>/dev/null
-  
-  # Check that download was correct
-  [ $( sha256sum boost_1_60_0.tar.bz2 | cut -d ' ' -f 1 ) == \
-    "686affff989ac2488f79a97b9479efb9f2abae035b5ed4d8226de6857933fd3b" ] \
-    || ( prfalse 'Corrupt download' ; exit 1 )
+  # If boost sources are not present or not correct, download them
+  if [[ ! -f $WORKROOT/boost_1_60_0.tar.bz2 || ! $( sha256sum boost_1_60_0.tar.bz2 | cut -d ' ' -f 1 ) == \
+      "686affff989ac2488f79a97b9479efb9f2abae035b5ed4d8226de6857933fd3b" ]]; then
+    # Download Boost
+    prstat "Downloading boost 1.60..."
+    URL='http://sourceforge.net/projects/boost/files/boost/1.60.0/boost_1_60_0.tar.bz2/download'
+    wget -c "$URL" -O $WORKROOT/boost_1_60_0.tar.bz2 &>/dev/null
+    
+    # Check that download was correct
+    [ $( sha256sum boost_1_60_0.tar.bz2 | cut -d ' ' -f 1 ) == \
+      "686affff989ac2488f79a97b9479efb9f2abae035b5ed4d8226de6857933fd3b" ] \
+      || ( prfalse 'Corrupt download, retry build script' ; exit 1 )
+  fi
 
-  # Copy source tarball into chroot
-  cp $WORKROOT/boost_1_60_0.tar.bz2 $WORKROOT/ubuntu14/
+  # If tarball is not in chroot, copy it in, and remove build completion flag if present
+  if [[ ! -f $WORKROOT/ubuntu14/boost_1_60_0.tar.bz2 ]]; then
+    # Copy source tarball into chroot
+    cp $WORKROOT/boost_1_60_0.tar.bz2 $WORKROOT/ubuntu14/
+    rm $WORKROOT/ubuntu14/.boost
+  fi
+
   # Mark process complete
   touch $WORKROOT/.boost
   prtrue "Completed download of boost and placed into chroot"
@@ -229,29 +261,55 @@ else
   prtrue "Already have Boost downloaded and copied into chroot"
 fi 
 
-# We assume that every build the repo is changed (for development reasons)
-# so we just remove the old one and recopy it
-sudo rm -rf $WORKROOT/ubuntu14/calibrd
+# If user has added clean launch parameters, completely remove the repo within the chroot
+if [[ $1 = "clean" ]]; then
+  prstat "User has requested to clean repository in chroot"
+  sudo rm -rf $WORKROOT/ubuntu14/calibrd
+fi
 
-# Copy the calibrd repository
-prstat "Copying $BINARYNAME Git repository..."
-cp -rf $REPODIR $WORKROOT/ubuntu14/
+# Copy all updated files in the repository to the work directory
+prstat "Updating $BINARYNAME Git repository..."
+cp -rfup $REPODIR $WORKROOT/ubuntu14/
 
-# Bind mount system folders and mark that procedure was started (so it can be cleaned up)
-prstat "Mounting system folders inside chroot"
-MOUNTEDSYSFOLDERS="1"
-sudo mount -o bind /dev $WORKROOT/ubuntu14/dev
-sudo mount -o bind /dev/pts $WORKROOT/ubuntu14/dev/pts
-sudo mount -o bind /sys $WORKROOT/ubuntu14/sys
-sudo mount -o bind /proc $WORKROOT/ubuntu14/proc
-sudo cp /etc/resolv.conf $WORKROOT/ubuntu14/etc/resolv.conf
+# Test whether anything in the repo have been changed
+# If there is change, unset the complete marker inside the repo in the chroot
+# Code taken from https://www.jpablo128.com/how-to-detect-changes-in-a-directory-with-bash/
+DIR_TO_CHECK="$REPODIR"
+PATH_TO_EXCLUDE="$REPODIR/{*.md}"
 
-# Copy chroot build script into chroot and run it
-prstat "Starting chrooted build script"
-sudo chroot $WORKROOT/ubuntu14 bash /calibrd/buildcalibrd.sh
+OLD_SUM_FILE="$HOME/calibrd_stat.txt"
 
-prstat "Copying out completed steemd, which will run on any version of ubuntu from 14.04 to 17.04"
-cp $WORKROOT/ubuntu14/calibrd/build/programs/steemd/steemd $REPODIR/..
+# If this is first run, create an empty stat file
+if [[ -e $OLD_SUM_FILE ]]
+then OLD_SUM="$(cat $OLD_SUM_FILE)"
+else OLD_SUM="nothing"
+fi
 
-cleanup
+NEW_SUM=`find $DIR_TO_CHECK/* \! -path "$PATH_TO_EXCLUDE"  -print0| xargs -0 du -b --time --exclude=$PATH_TO_EXCLUDE | sort -k4,4 | sha1sum | awk '{print $1}'`
+
+if [[ "$OLD_SUM" != "$NEW_SUM" || $1 = "rebuild" || ! -f $WORKROOT/ubuntu14/calibrd/programs/steemd/steemd ]]; then
+  prstat "Stuff has changed. Running build"
+  # Remove build completed flag
+  if [[ -f $WORKROOT/ubuntu14/calibrd/build/.calibrd ]]; then rm -f $WORKROOT/ubuntu14/calibrd/build/.calibrd
+  fi
+
+  # Bind mount system folders and mark that procedure was started (so it can be cleaned up)
+  prstat "Mounting system folders inside chroot"
+  MOUNTEDSYSFOLDERS="1"
+  sudo mount -o bind /dev $WORKROOT/ubuntu14/dev
+  sudo mount -o bind /dev/pts $WORKROOT/ubuntu14/dev/pts
+  sudo mount -o bind /sys $WORKROOT/ubuntu14/sys
+  sudo mount -o bind /proc $WORKROOT/ubuntu14/proc
+  sudo cp /etc/resolv.conf $WORKROOT/ubuntu14/etc/resolv.conf
+
+  # Copy chroot build script into chroot and run it
+  prstat "Starting chrooted build script"
+  sudo chroot $WORKROOT/ubuntu14 bash /calibrd/buildcalibrd.sh $1
+
+  echo $NEW_SUM > $OLD_SUM_FILE
+  cleanup
+else
+  prtrue "No changes in $REPODIR, nothing to do"
+fi
+
 # The End
